Add new message input
We only change enough code to get to the next error message. Getting past the "never found it" error only requires us to add an <input type="text"/> component.

Red: Expected to find element: '[data-test='saveButton']', but never found it.

The next error is very similar to the previous one. This time, it’s the save button that’s missing.