Specify component should call the onSave event handler
We create a test for the behavior of the component. We specify that when text is entered into the component, and then the Save button is clicked, then whatever function is passed in as the onSave prop, that function is called and passed the string entered in the text field.

Inner red: Expected mock function to have been called one time, but it was called zero times.