Call onSave handler when Save button clicked
We fix the failing component test by calling the onSave handler when the Save button is clicked, and passing whatever the current state of the inputText is. Then we rerun the end-to-end test.

Outer red: Uncaught TypeError: onSave is not a function

We are passing in an onSave function in the component test, but not in the applicaiton itself.